fix: remove dead crypto code and fix showPubkey type (4→2 errors)
Removed unused code:
- cryptography.ts: Deleted saveEncrypted/loadEncrypted functions (never called)
- cryptography.ts: Removed unused crypto import and algorithm constant

Type fix:
- showPubkey.ts: Cast publicKey to Uint8Array

Remaining 2 errors in fhe_test.ts (closed as not planned)
Effective production errors: 0

+++ src/libs/utils/showPubkey.ts
@@ -88,7 +88,7 @@ async function main() {
     const identity = await ucrypto.getIdentity(SIGNING_ALGORITHM)
 
     // Get the public key
-    const publicKeyHex = uint8ArrayToHex(identity.publicKey)
+    const publicKeyHex = uint8ArrayToHex(identity.publicKey as Uint8Array)
 
     // Output to file if -o flag provided, otherwise display to console
     if (outputFile) {
